chore(quality): split onboarding stage files, enforce components-per-file
Two onboarding stage files exported 3 React components each (limit is 2):
- Chat.tsx -> move OnboardingChatStream into ChatStream.tsx
- RevealTodos.tsx -> move SelectedTodoIndicator into SelectedTodoIndicator.tsx

Import sites updated (stages/index.ts barrel, CompletedStagesTimeline,
demo/page.tsx). Public exports unchanged. Verified:
check-components-per-file.mjs -> all files within limit.

Enforce the lane via marker file .github/quality-gate/enforced/components-per-file
(no workflow edit -> cannot merge-conflict with other lane PRs).
