+++ .github/quality-gate/enforced/components-per-file
@@ -1,0 +1,1 @@
+enforced
