+++ apps/web/src/features/onboarding/components/stages/Chat.tsx
@@ -11,16 +11,10 @@ import * as m from "motion/react-m";
 import type { Dispatch } from "react";
 import { useEffect } from "react";
 import ChatBubbleBot from "@/features/chat/components/bubbles/bot/ChatBubbleBot";
-import ChatBubbleUser from "@/features/chat/components/bubbles/user/ChatBubbleUser";
-import { LoadingIndicator } from "@/features/chat/components/interface/LoadingIndicator";
-import {
-  BOT_BUBBLE_DEFAULTS,
-  USER_BUBBLE_DEFAULTS,
-} from "../../constants/bubbleDefaults";
+import { BOT_BUBBLE_DEFAULTS } from "../../constants/bubbleDefaults";
 import {
   MOTION_COMPOSER_CTA,
   MOTION_FADE_UP_LARGE,
-  MOTION_STREAM_MESSAGE,
 } from "../../constants/motion";
 import {
   type UseOnboardingChatReturn,
@@ -58,74 +52,6 @@ export function useChatStage(
   }, [pendingTodoMessage, todoDemoConvoId, dispatch]);
 
   return { welcome, todoDemo };
-}
-
-const RUN_NOW_PREFIX = "Execute this todo for me: ";
-
-export function OnboardingChatStream({
-  chat,
-  hideRunNowUserMessage = false,
-  hideBotAvatar = false,
-}: {
-  chat: UseOnboardingChatReturn;
-  hideRunNowUserMessage?: boolean;
-  hideBotAvatar?: boolean;
-}) {
-  const visibleMessages = hideRunNowUserMessage
-    ? chat.streamMessages.filter(
-        (msg) =>
-          !(
-            msg.role === "user" &&
-            typeof msg.content === "string" &&
-            msg.content.startsWith(RUN_NOW_PREFIX)
-          ),
-      )
-    : chat.streamMessages;
-  return (
-    <>
-      {visibleMessages.map((msg) => (
-        <m.div key={msg.id} {...MOTION_STREAM_MESSAGE}>
-          {msg.role === "user" ? (
-            <ChatBubbleUser
-              {...USER_BUBBLE_DEFAULTS}
-              text={msg.content}
-              message_id={msg.id}
-              date={msg.createdAt.toISOString()}
-              fileData={msg.fileData}
-            />
-          ) : (
-            <ChatBubbleBot
-              {...BOT_BUBBLE_DEFAULTS}
-              text={msg.content}
-              message_id={msg.id}
-              loading={msg.status === "sending"}
-              tool_data={msg.tool_data ?? undefined}
-              todo_progress={msg.todo_progress ?? undefined}
-              memory_data={msg.memory_data ?? undefined}
-              image_data={msg.image_data ?? undefined}
-              date={msg.createdAt.toISOString()}
-              hideAvatar={hideBotAvatar}
-            />
-          )}
-        </m.div>
-      ))}
-
-      {chat.isChatSending &&
-        !visibleMessages.some(
-          (msg) => msg.role === "assistant" && msg.content,
-        ) && (
-          <LoadingIndicator
-            loadingText={
-              hideRunNowUserMessage
-                ? "Auto-executing todo…"
-                : "GAIA is thinking…"
-            }
-            loadingTextKey={0}
-            noPadding={hideRunNowUserMessage}
-          />
-        )}
-    </>
-  );
 }
 
 export function Chat({ state }: Omit<ChatProps, "dispatch" | "chat">) {
